// frontnd/src/interfaces/coins.ts
// Generated by https://quicktype.io

export interface Coins {
    _id:        string;
    transac_id: number;
    symbol:     Symbol;
    side:       Side;
    price:      number;
    date:       string;
    __v:        number;
}

export enum Side {
    Buy = "Buy",
    Sell = "Sell",
}

export enum Symbol {
    Xbtusd = "XBTUSD",
}